[REFACTOR] Add rev field back to KeyPair for Builder
Builder accesses the `rev` field directly, unfortunately. We add the
field back here as a `String` type, also adding a polymorphic `rev`
function that can be used instead, as a way to smooth the transition
to `KeyRevision` types in Builder.

However, I suspect that stronger use of types in Builder may
ultimately make this function call obsolete.

Signed-off-by: Christopher Maier <cmaier@chef.io>