[REFACTOR] Rename SymKey to RingKey
A "ring key" is the only kind of symmetric key we have in the system,
and it makes more sense to refer to the thing by the role it plays in
the system. The fact that it happens to be a symmetric key is less
important for understanding how we use it.

Signed-off-by: Christopher Maier <cmaier@chef.io>