[REFACTOR] Helper macro for TryFrom<PathBuf> implementations
We'll want all our keys to be able to be rehydrated from files, and
all that logic will be the same, so let's implement a helper macro.

Depending on how things ultimately shake out, we may be able to
replace this macro with a blanket implementation that covers all our
keys.

Signed-off-by: Christopher Maier <cmaier@chef.io>