[REFACTOR] Another Great Collapsing
This changes up a fair amount, but eliminates quite a bit.

We now have a `KeyFile` trait that consolidates all the stuff needed
for writing a key file to disk (e.g., extension, permissions, etc.).

We also generate all the keys using a new `gen_key!` macro, which
creates the types (all are the same structure, after all), and
implements several traits that make them behave as keys.

Signed-off-by: Christopher Maier <cmaier@chef.io>